Move `scala.interal.Chars` to `dotc.util`
This class was there for a single use of `digit2int` in `FromDigits`.
`digit2int` was copied inside of `FromDigit`.
