+++ compiler/src/dotty/tools/dotc/core/NameOps.scala
@@ -5,7 +5,7 @@ import java.security.MessageDigest
 import scala.io.Codec
 import Int.MaxValue
 import Names._, StdNames._, Contexts._, Symbols._, Flags._, NameKinds._, Types._
-import scala.internal.Chars.{isOperatorPart, digit2int}
+import util.Chars.{isOperatorPart, digit2int}
 import Definitions._
 import nme._
 import Decorators.concat
