+++ compiler/src/dotty/tools/dotc/util/Chars.scala
@@ -1,4 +1,4 @@
-package scala.internal
+package dotty.tools.dotc.util
 
 import scala.annotation.switch
 import java.lang.{Character => JCharacter}
