+++ library/src/scala/util/FromDigits.scala
@@ -1,7 +1,6 @@
 package scala.util
 import scala.math.{BigInt}
 import quoted._
-import internal.Chars.digit2int
 import annotation.internal.sharable
 
 /** A type class for types that admit numeric literals.
@@ -91,6 +90,19 @@ object FromDigits {
       i += 1
     }
     if (negated) -value else value
+  }
+
+  /** Convert a character digit to an Int according to given base,
+   *  -1 if no success
+   */
+  private def digit2int(ch: Char, base: Int): Int = {
+    val num = (
+      if (ch <= '9') ch - '0'
+      else if ('a' <= ch && ch <= 'z') ch - 'a' + 10
+      else if ('A' <= ch && ch <= 'Z') ch - 'A' + 10
+      else -1
+      )
+    if (0 <= num && num < base) num else -1
   }
 
   /** Convert digit string to Int number
